+++ tests/Functional/CacheInvalidatorTest.php
@@ -21,14 +21,20 @@ class CacheInvalidatorTest extends VarnishTestCase
 {
     public function testInvalidateTags()
     {
+        if (getenv('VARNISH_MODULES_VERSION')) {
+            $uri = '/tags_xkey.php';
+        } else {
+            $uri = '/tags.php';
+        }
+
         $cacheInvalidator = new CacheInvalidator($this->getProxyClient());
 
-        $this->assertMiss($this->getResponse('/tags.php'));
-        $this->assertHit($this->getResponse('/tags.php'));
+        $this->assertMiss($this->getResponse($uri));
+        $this->assertHit($this->getResponse($uri));
 
         $cacheInvalidator->invalidateTags(['tag1']);
         $cacheInvalidator->flush();
 
-        $this->assertMiss($this->getResponse('/tags.php'));
+        $this->assertMiss($this->getResponse($uri));
     }
 }