+++ CHANGELOG.md
@@ -3,6 +3,15 @@ Changelog
 
 See also the [GitHub releases page](https://github.com/FriendsOfSymfony/FOSHttpCache/releases).
 
+2.2.0 (unreleased)
+------------------
+
+### Varnish
+
+* Added support for the more efficient xkey cache tag system. BAN remains the
+  default cache tagging system, but if you can install the varnish modules in
+  your system, it is recommended to update to xkey.
+
 2.1.1
 -----